pyproject.toml
[project]
name = "context-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "django>=5.2.8",
    "django-ninja>=1.5.0",
    "django-ninja-extra>=0.30.6",
    "dlt[parquet]>=1.18.2",
    "dotenv>=0.9.9",
    "langchain-chroma>=1.0.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "prefect>=3.6.4",
    "psycopg2-binary>=2.9.11",
]
